location.jsに切り出した
